+++ src/components/service-menu.vue
@@ -17,9 +17,9 @@
         <div v-else class="item-disabled">{{ $t("stop_service") }}</div>
         <div v-if="!locked && running" v-on:click.stop="control('restart')" class="item">{{ $t("restart_service") }}</div>
         <div v-else class="item-disabled">{{ $t("restart_service") }}</div>
-        <div class="item-seperator"></div>
-        <div v-if="!locked" v-on:click.stop="reboot()" class="item">{{ $t("reboot_device") }}</div>
-        <div v-else class="item-disabled">{{ $t("reboot_device") }}</div>
+        <div v-if="!$server.docker" class="item-seperator"></div>
+        <div v-if="!$server.docker && !locked" v-on:click.stop="reboot()" class="item">{{ $t("reboot_device") }}</div>
+        <div v-else-if="!$server.docker" class="item-disabled">{{ $t("reboot_device") }}</div>
         <div class="item-seperator"></div>
         <div class="item" v-on:click="about">{{ $t("about") }}</div>
         <router-link to="/config/interface" class="item">{{ $t("config") }}</router-link>
@@ -49,9 +49,9 @@
         <div v-else class="item-disabled">{{ $t("stop_service") }}</div>
         <div v-if="!locked && running" v-on:click.stop="control('restart')" class="item">{{ $t("restart_service") }}</div>
         <div v-else class="item-disabled">{{ $t("restart_service") }}</div>
-        <div class="item-seperator"></div>
-        <div v-if="!locked" v-on:click.stop="reboot()" class="item">{{ $t("reboot_device") }}</div>
-        <div v-else class="item-disabled">{{ $t("reboot_device") }}</div>
+        <div v-if="!$server.docker" class="item-seperator"></div>
+        <div v-if="!$server.docker && !locked" v-on:click.stop="reboot()" class="item">{{ $t("reboot_device") }}</div>
+        <div v-else-if="!$server.docker" class="item-disabled">{{ $t("reboot_device") }}</div>
         <div class="item-seperator"></div>
         <div class="item" v-on:click="about">{{ $t("about") }}</div>
         <router-link to="/config/interface" class="item">{{ $t("config") }}</router-link>