+++ src/views/terminal.vue
@@ -5,7 +5,7 @@
             <div v-for="(section, title) in info" :key="title">
                 <router-link :to="`/system/${title}`">{{ translate(title) }}</router-link>
             </div>
-            <router-link to="/system/filesystem">{{ translate("file_system") }}</router-link>
+            <router-link v-if="!$server.docker" to="/system/filesystem">{{ translate("file_system") }}</router-link>
             <router-link v-if="temp && (temp || {}).main >= 0" to="/system/temp">{{ translate("temperature") }}</router-link>
             <router-link to="/system/terminal" class="active">{{ $t("terminal") }}</router-link>
         </div>