+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 At HOOBS we are always working to make the software better. Here are the notiable changes in this version.
 
 * Auto close notifications
+* Added support for Docker containers
 
 ## 3.1.19
 At HOOBS we are always working to make the software better. Here are the notiable changes in this version.